+++ src/rapido/plone/tests/resources/rapido/testapp/blocks/knowledge.py
@@ -1,0 +1,2 @@
+def jokes(context):
+    return ['France is bacon', 'Francis Bacon', ]