+++ src/rapido/plone/tests/resources/rapido/testapp/blocks/call.py
@@ -1,0 +1,2 @@
+def call_me(context):
+    return context.params['x'] * context.params['y']