+++ src/rapido/plone/tests/resources/rapido/testapp/blocks/otherapp.py
@@ -1,0 +1,5 @@
+def create_external_record(context):
+    other = context.rapido('otherapp')
+    record = other.create_record()
+    record['something'] = 12
+    return len(other.records())