+++ CHANGES.rst
@@ -8,6 +8,9 @@ Changelog
 - Ajax links
   [ebrehault]
 
+- Better context.content computing and reinjection
+  [ebrehault]
+
 
 1.0 (2015-11-17)
 ----------------