+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.0.2 (unreleased)
+1.0.2 (2016-04-09)
 ------------------
 
 - Add rapidoLoad Javascript event